refactor: Change sampleId type from Int to String
This commit updates the `getUriFromSampleId` function in `NeptuneMediaPlayer` to accept a `String` instead of an `Int` for the `sampleId`. This change makes the function more robust by using the `hashCode()` of the string to determine which sample audio to play.

Key changes:
- Modified `getUriFromSampleId` to take a `String` parameter and use its hash code for record selection.
- Updated calls to this function in `MainScreen.kt` and `PostScreen.kt` to pass the string `sample.id` directly, removing the `toInt()` conversion.
- Adjusted the corresponding unit test in `NeptuneMediaPlayerTest.kt` to reflect the new parameter type.

+++ app/src/main/java/com/neptune/neptune/ui/post/PostScreen.kt
@@ -188,8 +188,7 @@ fun PostScreen(
                           .clickable(
                               onClick = {
                                 mediaPlayer.togglePlay(
-                                    mediaPlayer.getUriFromSampleId(
-                                        uiState.sample.id.toInt())) // TODO CHANGE THIS
+                                    mediaPlayer.getUriFromSampleId(uiState.sample.id))
                               })
                           .aspectRatio(1.6f)
                           .border(1.dp, NepTuneTheme.colors.onBackground, RoundedCornerShape(8.dp))
